requirements.txt
# Core FastAPI and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.35.0
python-multipart==0.0.20

# Database and ORM
sqlalchemy==2.0.41
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.1

# AI and ML
google-generativeai==0.8.3
langchain==0.3.12
langchain-google-genai==2.0.8
langchain-community==0.3.12

# Document Processing
markitdown==0.0.1a2
pypdfium2==4.30.1
python-docx==1.2.0
python-pptx==1.0.2

# Data Processing and Analysis
pandas==2.3.1
numpy==2.2.6
matplotlib==3.10.3
plotly==5.24.1
seaborn==0.13.2

# Utilities
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.4
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
structlog==24.5.0
prometheus-client==0.21.1
sentry-sdk[fastapi]==2.19.2

# Caching and Performance
redis-py==5.2.1
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pytest-cov==6.0.0

# Development Tools
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Security
cryptography==44.0.0
bcrypt==4.2.1
